ntfy-handler/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
psycopg2-binary
requests
